v3: Allow listing of roles with GET /v3/roles
This endpoint fetches roles that a user has access to. Default order_by
value is 'created_at', rather than 'id' like it is for other tables,
since the roles resource is a combination of 7 tables with their own
auto-incrementing 'id' columns.

To represent this combination of 7 tables, we introduce a new Role model
backed by a dataset (a union of these tables), rather than by a table as
usual.

After the initial introduction of roles, there was an edge case where,
during a rolling deployment, roles created by older API instances would
not populate the table rows for these roles with 'guid's. This commit
handles that edge case by doing a 'just-in-time migration' before
returning the list of roles: it runs a quick query to generate a new
guid for any roles that do not have a guid already. In most cases, this
should be a no-op.

[Finishes #164176845]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Teal Stannard <tstannard@pivotal.io>
Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>