Revert "v3: Fix resource type for org quota delete jobs"
This commit did not fully fix the problem and introduced another
problem. After this change, we would get 500s when trying to get the job
by guid. The underlying error was something like:

```
Mysql2::Error: Table 'cloud_controller.organization_quota' doesn't exist
```

This is because there is a method `job.resource_exists?` which uses the
job's `resource_type` to construct query like `SELECT * FROM
<resource_type> WHERE guid = <guid>`. But since `organization_quota` is
not the name of the table, this was failing.

Going to try a different approach to fix this problem.

[#164973539]

This reverts commit 7342b34cf39c5ebf297d7a73fd1f407b17f328a2.