V3 `POST /v3/routes` has helpful error messages
...when interacting with routing API.

The error handling is nuanced. For example, on the lowest layer a
`RoutingApi::UaaUnavailable` is thrown, but that's not what the
controller receives; it receives a massaged error, which we match
against and turn that into yet a different error. Same error, three
different manifestations.

We appeased Rubocop's `Cyclomatic Complexity` demands by moving the
controllers error into its own, separate method, which we feel _reduces_
the readability of the code slightly, but that's a small price to pay.

[#169279023]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Teal Stannard <tstannard@pivotal.io>