v3(services): refactor delete for service instance
We have moved to a pattern where a controller either uses an action to
do something, or schedules a job to do something, and the job uses an
action under the hood. The services instances delete endpoint was using
an action to schedule a job, which could be confusing as it's different
to other endpoints.

[#171728591](https://www.pivotaltracker.com/story/show/171728591)